feat: feed 수정

+++ src/main/java/com/lec/spring/mytrip/domain/PostAttachment.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class PostAttachment {
-    private Long boardAttachmentId;        // 첨부파일 ID
+    private int boardAttachmentId;        // 첨부파일 ID
     private Feed feed;                     // 피드
     private String boardAttachmentFile;    // 저장된 파일명
     private String filepath;               // 파일 경로
